AUTH: fix refresh errors
This fixes two problems: first we want to ignore errors when we are adding extra
context to our sentry reporting. This was leading to 500 responses that should
have been 4XX. Second, we want to store the refresh token in cookies when the
serializer passes validation - this was leading to error messages being stored as
the refresh token.