FRONTEND: check for redirect url in response rather than query parameters
Previously, we were checking for the presence of a next value using the
value of next in the query parameters. It is possible, however,  for
the server to supply a next value even when it does not receive one
(using the DEFAULT_REDIRECT_URL from settings). This meant that the
frontend was ignoring the default value in the case where the next
parameter was missing.